+++ app/main/shifts-application/page.tsx
@@ -1,11 +1,16 @@
-import { ShiftsApplyTable } from "./shifts_apply_table";
+'use client'
+import React from 'react';
+import ShiftsApplyTable from './shifts_apply_table';
+import { useAuth } from '@/providers/UserContextProvider';
+import { Employee } from '@/types/class.service';
 
 
-export default async function page() {
-   
-  return (
+export default function Page() {
+
+  const { user } = useAuth<Employee>();
+  return user && (
     <>
-      <ShiftsApplyTable />
+      <ShiftsApplyTable user={user} />
     </>
-  )
+  );
 }
